refactor: remove old code
