--- libraries/cfg-core/lib/nodes/Node.ts
@@ -1,41 +1,0 @@
-/*
- * Copyright 2020-2021 Delft University of Technology and SynTest contributors
- *
- * This file is part of SynTest Framework - SynTest Core.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * Interface for a Node.
- *
- * @author Dimitri Stallenberg
- */
-export interface Node {
-  readonly type: NodeType;
-
-  readonly id: string;
-
-  readonly lines: readonly number[];
-  readonly statements: string[];
-
-  readonly description?: string;
-}
-
-export enum NodeType {
-  Intermediary = "intermediary",
-  Branch = "branch",
-  Placeholder = "placeholder",
-  Root = "root",
-  Normal = "normal",
-}
